requirements.txt
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
decorator==4.4.2
Flask==1.1.2
future==0.18.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
newsapi-python==0.2.6
oauthlib==3.1.0
pycparser==2.20
PySocks==1.7.1
python-telegram-bot==12.7
requests==2.23.0
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.14.0
tornado==6.0.4
tweepy==3.8.0
urllib3==1.25.9
Werkzeug==1.0.1
